docker/upload-to-dicom-services/requirements.txt
azure.storage.blob
azure.identity
requests